projekti

+++ Java-projektit/projekti 6/paikallinen_tallennus.js
@@ -31,6 +31,9 @@
 
   //Tyhjennetään paikallinen muisti
   function poista()
+  {
+    if (confirm("Oletko varma?"))
     {
       localStorage.clear();
     }
+  }
